setup osx applescript based osx app.
This adds a `starfix.app` which will be driven by a simple applescript (`starfix-urlhandler.applescript`).

There is a `build.sh` file to create the actual app - this build is better driven from maven/gradle but
for now this will do.

Script assumes `starfix` binary are bundled inside the bundle making it self-contained.

Installer should eventually do a symbolic link to `/usr/local/bin/starfix` but not done currently.